refactor and move out quicklook mapview to seperate file
