+++ Avenue.xcodeproj/xcuserdata/vincentneo.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -54,7 +54,7 @@
 		<key>GPXQuickLook.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>4</integer>
+			<integer>1</integer>
 		</dict>
 	</dict>
 	<key>SuppressBuildableAutocreation</key>
