+++ GPXQuickLook/PreviewViewController.swift
@@ -11,56 +11,13 @@ import Quartz
 import MapKit
 import CoreGPX
 
-class MapView: MKMapView {
-    var extent = GPXExtentCoordinates()
-    var length = 0.0
-    var timeInterval = 0.0
-    
-    func loadedGPXFile(_ root: GPXRoot) {
-        print("MapView: GPX Object Loaded \(root)")
-        for track in root.tracks {
-            for trackseg in track.tracksegments {
-                length += trackseg.length()
-            }
-            for trackseg in track.tracksegments {
-                guard let startTime = trackseg.trackpoints.first?.time,
-                    let endTime = trackseg.trackpoints.last?.time else { continue }
-                let timeBetween = endTime.timeIntervalSince(startTime)
-                timeInterval += timeBetween
-            }
-        }
-        
-        for waypoint in root.waypoints {
-            self.addAnnotation(waypoint)
-        }
-        
-        for track in root.tracks {
-            for segment in track.tracksegments {
-                let overlay = segment.overlay
-
-                self.addOverlay(overlay, level: .aboveLabels)
-                
-                for trkpt in segment.trackpoints {
-                    self.extent.extendAreaToIncludeLocation(trkpt.coordinate)
-                }
-            }
-        }
-        // reduce region span being too close to bounds of view
-        extent.region.span.latitudeDelta *= 1.25
-        extent.region.span.longitudeDelta *= 1.25
-        
-        self.setRegion(extent.region, animated: true)
-        
-    }
-}
-
 class PreviewViewController: NSViewController, QLPreviewingController, MKMapViewDelegate {
     
     @IBOutlet weak var elapsedView: NSVisualEffectView!
     @IBOutlet weak var distanceView: NSVisualEffectView!
     @IBOutlet weak var elapsedLabel: NSTextField!
     @IBOutlet weak var distanceLabel: NSTextField!
-    @IBOutlet weak var mapView: MapView!
+    @IBOutlet weak var mapView: QLMapView!
     
     enum PossibleErrors: Error {
         case fileIsNil
@@ -106,24 +63,32 @@ class PreviewViewController: NSViewController, QLPreviewingController, MKMapView
         
         // Call the completion handler so Quick Look knows that the preview is fully loaded.
         // Quick Look will display a loading spinner while the completion handler is not called.
+        
         guard let parser = GPXParser(withURL: url) else { handler(PossibleErrors.fileIsNil); return }
+        
         do {
             guard let gpx = try parser.fallibleParsedData(forceContinue: false) else { handler(PossibleErrors.fileIsNil); return }
             mapView.loadedGPXFile(gpx)
+            let locale = Locale.current
+            let useImperial = !locale.usesMetricSystem
+            
             let elapsedText = ElapsedTime.getString(from: mapView.timeInterval)
-            let distanceText = mapView.length.toDistance(useImperial: false)
+            let distanceText = mapView.length.toDistance(useImperial: useImperial)
+            
             if mapView.timeInterval == 0 {
                 elapsedLabel.isHidden = true
             }
             else {
                 elapsedLabel.stringValue = elapsedText
             }
+            
             distanceLabel.stringValue = distanceText
             
         }
         catch {
             handler(error)
         }
+        
         handler(nil)
     }
